// features/api/model/nozzleResource.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface NozzleResource {
  id: string;
  name: string;
  digits: number;
  current_reading: number;
  /** Relationships */
  organization_id: string;
  station_id: string;
  /** @nullable */
  station_name: string | null;
  tank_id: string;
  /** @nullable */
  tank_name: string | null;
  /** @nullable */
  product_name: string | null;
  /** Helpful for UI labels */
  created_at: string;
}