// features/api/model/storeNozzleRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface StoreNozzleRequest {
  /** @maxLength 255 */
  name: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  digits: number;
  /** e.g., 7-digit mechanical counter */
  station_id: string;
  /** Validate Tank exists AND belongs to the chosen Station */
  tank_id: string;
  /** @minimum 0 */
  current_reading: number;
}