// features/api/model/updateNozzleRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface UpdateNozzleRequest {
  /** @maxLength 255 */
  name?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  digits?: number;
  station_id?: string;
  /** If updating tank, ensure it matches the station (use new station_id if provided, else current model's) */
  tank_id?: string;
  /** @minimum 0 */
  current_reading?: number;
}