// features/api/model/stationsIndex200.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { StationResource } from './stationResource';
import type { StationsIndex200Links } from './stationsIndex200Links';
import type { StationsIndex200Meta } from './stationsIndex200Meta';

export type StationsIndex200 = {
  data: StationResource[];
  links: StationsIndex200Links;
  meta: StationsIndex200Meta;
};