// features/api/model/tanksIndex200.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { TankResource } from './tankResource';
import type { TanksIndex200Links } from './tanksIndex200Links';
import type { TanksIndex200Meta } from './tanksIndex200Meta';

export type TanksIndex200 = {
  data: TankResource[];
  links: TanksIndex200Links;
  meta: TanksIndex200Meta;
};